+++ src/components/card/card-types.d.ts
@@ -1,1 +1,0 @@
-export type CardSizes = "small" | "medium" | "large";