+++ src/components/card/card.config.js
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const CARD_SIZES = ["small", "medium", "large"];
-export const CARD_ALIGNS = ["center", "left", "right"];