+++ src/components/card/index.ts
@@ -1,0 +1,8 @@
+export { default, default as Card } from "./card.component";
+export type { CardProps } from "./card.component";
+export { default as CardColumn } from "./card-column";
+export type { CardColumnProps } from "./card-column";
+export { default as CardFooter } from "./card-footer";
+export type { CardFooterProps } from "./card-footer";
+export { default as CardRow } from "./card-row";
+export type { CardRowProps } from "./card-row";