+++ src/components/image/image.style.ts
@@ -13,19 +13,24 @@ export interface StyledImageProps
   extends MarginProps,
     BackgroundProps,
     LayoutProps {
-  /** alt property to display when an img fails to load */
+  /** HTML alt property to display when an img fails to load */
   alt?: string;
-  /** any valid file path, passing this will render the component as an img element */
+  /** Any valid file path, passing this will render the component as an img element */
   src?: string;
+  /** HTML hidden property to indicate whether to remain hidden visually and from screen readers */
+  hidden?: boolean;
   /** Children elements, will only render if component is a div element */
   children?: React.ReactNode;
 }
 
-const StyledImage = styled.div.attrs(({ src, children }: StyledImageProps) => ({
-  ...(src && { as: "img" }),
-  children: src ? undefined : children,
-  src,
-}))<StyledImageProps>`
+const StyledImage = styled.div.attrs(
+  ({ src, children, hidden = false }: StyledImageProps) => ({
+    ...(src && { as: "img" }),
+    children: src ? undefined : children,
+    src,
+    hidden,
+  })
+)<StyledImageProps>`
   ${margin}
   ${layout}